[Firmware] Fix data size of some configuration items and messages

+++ firmware/destroyer.h
@@ -5,11 +5,11 @@
 #include <stdbool.h>
 
 typedef struct destroyer_data_s {
-    int16_t I_limit; // in mA  // AAMMM
+    uint16_t I_limit; // in mA  // AAMMM
     int16_t count;
     bool dut_is_active; // Actual current value
     bool sel_to_manage;
-    uint8_t avg_mode;   // See INA233 datasheet, only 3 bit significative
+    uint16_t avg_mode;   // See INA233 datasheet, only 3 bit significative
     uint8_t out_status; // 0x00 - Always OFF, 0xEE - always ON, 0xAA - auto
     uint16_t T_hold_us; // in 100 us // SMMMU 6.5535
